+++ Source/Core/HCAssertThat.h
@@ -38,54 +38,6 @@ FOUNDATION_EXPORT void HC_assertThatWithLocation(id testCase, id actual, id <HCM
 #endif
 
 
-typedef id (^HCAssertThatAfterActualBlock)() __attribute__((deprecated));
-
-OBJC_EXPORT void HC_assertThatAfterWithLocation(id testCase, NSTimeInterval maxTime,
-                                                HCAssertThatAfterActualBlock actualBlock,
-                                                id <HCMatcher> matcher,
-                                                const char *fileName, int lineNumber) __attribute__((deprecated));
-
-#define HC_assertThatAfter(maxTime, actualBlock, matcher)  \
-    HC_assertThatAfterWithLocation(self, maxTime, actualBlock, matcher, __FILE__, __LINE__)
-
-#define HC_futureValueOf(actual) ^{ return actual; }
-
-#ifdef HC_SHORTHAND
-/*!
- * @abstract assertThatAfter(maxTime, actualBlock, matcher) -
- * Asserts that a value provided by a block will satisfy matcher in less than the specified time.
- * @param maxTime Max time (in seconds) in which the matcher has to be satisfied.
- * @param actualBlock A block providing the object to evaluate until timeout or the matcher is satisfied.
- * @param matcher The matcher to satisfy as the expected condition.
- * @deprecated Version 4.2.0. Use @ref assertWithTimeout instead.
- * @discussion assertThatAfter checks several times if the matcher is satisfied before timeout. To
- * evaluate the matcher, the <em>actualBlock</em> will provide updated values of actual. If the
- * matcher is not satisfied after <em>maxTime</em>, it is reported to the @ref HCTestFailureReporterChain.
- *
- * An easy way of defining the actualBlock is using the macro <code>futureValueOf(actual)</code>,
- * which also improves readability.
- *
- * <b>Name Clash</b><br />
- * In the event of a name clash, don't <code>#define HC_SHORTHAND</code> and use the synonym
- * HC_assertThatAfter instead.
-*/
-#define assertThatAfter HC_assertThatAfter
-
-/*!
- * @abstract futureValueOf(actual) -
- * Evaluates actual value at future time.
- * @param actual The object to evaluate as the actual value.
- * @deprecated Version 4.2.0. Use @ref thatEventually instead.
- * @discussion Wraps <em>actual</em> in a block so that it can be repeatedly evaluated by @ref assertThatAfter.
- *
- * <b>Name Clash</b><br />
- * In the event of a name clash, don't <code>#define HC_SHORTHAND</code> and use the synonym
- * HC_futureValueOf instead.
- */
-#define futureValueOf HC_futureValueOf
-#endif
-
-
 typedef id (^HCFutureValue)();
 
 OBJC_EXPORT void HC_assertWithTimeoutAndLocation(id testCase, NSTimeInterval timeout,