+++ Source/Core/HCAssertThat.m
@@ -39,13 +39,6 @@ void HC_assertThatWithLocation(id testCase, id actual, id <HCMatcher> matcher,
         reportMismatch(testCase, actual, matcher, fileName, lineNumber);
 }
 
-void HC_assertThatAfterWithLocation(id testCase, NSTimeInterval maxTime,
-                                    HCAssertThatAfterActualBlock actualBlock, id <HCMatcher> matcher,
-                                    const char *fileName, int lineNumber)
-{
-    HC_assertWithTimeoutAndLocation(testCase, maxTime, actualBlock, matcher, fileName, lineNumber);
-}
-
 void HC_assertWithTimeoutAndLocation(id testCase, NSTimeInterval timeout, HCFutureValue actualBlock, id <HCMatcher> matcher, const char *fileName, int lineNumber)
 {
     BOOL match;